refactor: tela cliente
