+++ app/src/main/java/com/anotasmart/ui/components/AppTabRow.kt
@@ -1,0 +1,41 @@
+package com.anotasmart.ui.components
+
+import androidx.compose.material3.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.vector.ImageVector
+
+@Composable
+fun AppTabRow(
+    tabIndex: Int,
+    tabs: List<String>,
+    icons: List<ImageVector>,
+    onTabSelected: (Int) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    PrimaryTabRow(
+        selectedTabIndex = tabIndex,
+        containerColor = MaterialTheme.colorScheme.surface,
+        contentColor = MaterialTheme.colorScheme.primary,
+        modifier = modifier
+    ) {
+        tabs.forEachIndexed { index, title ->
+            Tab(
+                selected = tabIndex == index,
+                onClick = { onTabSelected(index) },
+                text = {
+                    Text(
+                        text = title,
+                        style = MaterialTheme.typography.titleSmall
+                    )
+                },
+                icon = {
+                    Icon(
+                        imageVector = icons[index],
+                        contentDescription = null
+                    )
+                }
+            )
+        }
+    }
+}
